fix: Update ubuntu install scripts

+++ ubuntu-openjdk.sh
@@ -1,6 +1,7 @@
 #! /bin/bash
 
-sudo apt install -y dirmngr
-sudo add-apt-repository -y ppa:openjdk-r/ppa
-sudo apt update -y
-sudo apt install -y default-jdk-headless
+if ! [ -e /etc/apt/sources.list.d/openjdk.list ]; then
+    sudo apt install -y dirmngr
+    sudo add-apt-repository -y ppa:openjdk-r/ppa
+    sudo apt install -y default-jdk-headless
+fi
